# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ALT Agency Automation Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /leads:
    get:
      operationId: listLeads
      tags: [leads]
      summary: List all leads
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: niche
          in: query
          schema:
            type: string
        - name: location
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
        - name: offset
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of leads
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Lead"
    post:
      operationId: createLead
      tags: [leads]
      summary: Create a new lead
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LeadInput"
      responses:
        "201":
          description: Created lead
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
  /leads/{id}:
    get:
      operationId: getLead
      tags: [leads]
      summary: Get lead by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Lead
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
    patch:
      operationId: updateLead
      tags: [leads]
      summary: Update lead
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LeadUpdate"
      responses:
        "200":
          description: Updated lead
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
    delete:
      operationId: deleteLead
      tags: [leads]
      summary: Delete a lead
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /leads/find:
    post:
      operationId: findLeads
      tags: [leads]
      summary: AI-powered lead discovery
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LeadFinderInput"
      responses:
        "200":
          description: Discovered leads
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Lead"
...
  /campaigns:
    get:
      operationId: listCampaigns
      tags: [campaigns]
      summary: List all outreach campaigns
      responses:
        "200":
          description: List of campaigns
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Campaign"
    post:
      operationId: createCampaign
      tags: [campaigns]
      summary: Create a new campaign
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CampaignInput"
      responses:
        "201":
          description: Created campaign
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
  /campaigns/{id}:
    get:
      operationId: getCampaign
      tags: [campaigns]
      summary: Get campaign by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Campaign
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
    patch:
      operationId: updateCampaign
      tags: [campaigns]
      summary: Update campaign
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CampaignUpdate"
      responses:
        "200":
          description: Updated campaign
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
    delete:
      operationId: deleteCampaign
      tags: [campaigns]
      summary: Delete a campaign
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /campaigns/{id}/launch:
    post:
      operationId: launchCampaign
      tags: [campaigns]
      summary: Launch an outreach campaign
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Campaign launched
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
  /campaigns/{id}/emails:
    get:
      operationId: listCampaignEmails
      tags: [campaigns]
      summary: List emails in a campaign
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Campaign emails
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CampaignEmail"
  /campaigns/generate-email:
    post:
      operationId: generateOutreachEmail
      tags: [campaigns]
      summary: AI-generate a personalized outreach email
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmailGenerateInput"
      responses:
        "200":
          description: Generated email
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GeneratedEmail"

  # ── Analytics ──────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Lead:
      type: object
      required: [id, workspaceId, businessName, status, createdAt]
      properties:
        id:
          type: integer
        workspaceId:
          type: integer
        businessName:
          type: string
        website:
          type: ["string", "null"]
        contactEmail:
          type: ["string", "null"]
        contactName:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
        niche:
          type: ["string", "null"]
        status:
          type: string
        auditScore:
          type: ["integer", "null"]
        auditId:
          type: ["integer", "null"]
        tags:
          type: array
          items:
            type: string
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
    LeadInput:
      type: object
      required: [businessName]
      properties:
        businessName:
          type: string
        website:
          type: string
        contactEmail:
          type: string
        contactName:
          type: string
        phone:
          type: string
        location:
          type: string
        niche:
          type: string
        notes:
          type: string
    LeadUpdate:
      type: object
      properties:
        businessName:
          type: string
        status:
          type: string
        contactEmail:
          type: string
        contactName:
          type: string
        notes:
          type: string
        tags:
          type: array
          items:
            type: string
    LeadFinderInput:
      type: object
      required: [niche, location]
      properties:
        niche:
          type: string
        location:
          type: string
        count:
          type: integer
...
    Campaign:
      type: object
      required: [id, workspaceId, name, status, type, createdAt]
      properties:
        id:
          type: integer
        workspaceId:
          type: integer
        name:
          type: string
        status:
          type: string
        type:
          type: string
        subject:
          type: ["string", "null"]
        body:
          type: ["string", "null"]
        sentCount:
          type: integer
        openCount:
          type: integer
        replyCount:
          type: integer
        openRate:
          type: number
        replyRate:
          type: number
        scheduledAt:
          type: ["string", "null"]
        launchedAt:
          type: ["string", "null"]
        createdAt:
          type: string
    CampaignInput:
      type: object
      required: [name, type]
      properties:
        name:
          type: string
        type:
          type: string
        subject:
          type: string
        body:
          type: string
        scheduledAt:
          type: string
    CampaignUpdate:
      type: object
      properties:
        name:
          type: string
        subject:
          type: string
        body:
          type: string
        status:
          type: string
        scheduledAt:
          type: string
    CampaignEmail:
      type: object
      required: [id, campaignId, toEmail, status, createdAt]
      properties:
        id:
          type: integer
        campaignId:
          type: integer
        toEmail:
          type: string
        toName:
          type: ["string", "null"]
        subject:
          type: string
        status:
          type: string
        openedAt:
          type: ["string", "null"]
        repliedAt:
          type: ["string", "null"]
        createdAt:
          type: string
    EmailGenerateInput:
      type: object
      required: [leadId, type]
      properties:
        leadId:
          type: integer
        type:
          type: string
        tone:
          type: string
    GeneratedEmail:
      type: object
      required: [subject, body]
      properties:
        subject:
          type: string
        body:
          type: string

    # ── Analytics ─────────────────────────────────────────────────────────────
...